Add identity cmd and move sso, groups, users cmd inside of identity cmd
